pyproject.toml
[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "embedbase"
version = "1.1.9"
description = "Open-source API & SDK to integrate your data and easily hook them up to LLMs."
readme = "README.md"
authors = ["Different AI <louis@embedbase.xyz>"]
license = "MIT"
repository = "https://github.com/different-ai/embedbase"
homepage = "https://github.com/different-ai/embedbase"
keywords = ["embeddings", "machine learning", "artificial intelligence", "llm"]

classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
]

[tool.poetry.scripts]
# Entry points for the package
"embedbase" = "embedbase.__main__:run_app"

[tool.poetry.dependencies]
python = "^3.8"

fastapi = "^0.95.1"
uvicorn = {extras = ["standard"], version = "^0.15.0"}
pandas = "^1.3.4"
openai = "^0.27.0"
tenacity = "^8.0.1"
pydantic_yaml = {extras = ["pyyaml"], version = "^0.4.0"}
tiktoken = "^0.3.3"

[tool.poetry.dev-dependencies]
bandit = "^1.7.1"
black = {version = "^21.10b0", allow-prereleases = true}
darglint = "^1.8.1"
isort = {extras = ["colors"], version = "^5.10.1"}
mypy = "^0.910"
mypy-extensions = "^0.4.3"
pre-commit = "^2.15.0"
pydocstyle = "^6.1.1"
pylint = "^2.11.1"
pytest = "^7.3.1"
pyupgrade = "^2.29.1"
safety = "^1.10.3"
coverage = "^6.1.2"
coverage-badge = "^1.1.0"
pytest-html = "^3.1.1"
pytest-cov = "^4.0.0"
httpx = "^0.23.0"
pytest-asyncio = "^0.21.0"
requests-mock = "^1.10.0"
sentence-transformers = "^2.2.2"

# TODO: following integrations might be moved outside this repo - loose coupling
firebase-admin = "^6.1.0"
sentry-sdk = {extras = ["fastapi"], version = "^1.21.1"}
supabase = "^1.0.3"
psycopg = {extras = ["binary", "pool"], version = "^3.1.8"}
pgvector = "^0.1.6"

[tool.black]
target-version = ["py38"]
line-length = 88
color = true

exclude = '''
/(
    \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
    | env
    | venv
)/
'''

[tool.isort]
py_version = 38
line_length = 88
known_typing = ["typing", "types", "typing_extensions", "mypy", "mypy_extensions"]
sections = ["FUTURE", "TYPING", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]
include_trailing_comma = true
profile = "black"
multi_line_output = 3
indent = 4
color_output = true

[tool.mypy]
python_version = 3.8
pretty = true
show_traceback = true
color_output = true
allow_redefinition = false
check_untyped_defs = true
disallow_any_generics = true
disallow_incomplete_defs = true
ignore_missing_imports = true
implicit_reexport = false
no_implicit_optional = true
show_column_numbers = true
show_error_codes = true
show_error_context = true
strict_equality = true
strict_optional = true
warn_no_return = true
warn_redundant_casts = true
warn_return_any = true
warn_unreachable = true
warn_unused_configs = true
warn_unused_ignores = true

[tool.pytest.ini_options]
norecursedirs = ["hooks", "*.egg", ".eggs", "dist", "build", "docs", ".tox", ".git", "__pycache__"]
doctest_optionflags = ["NUMBER", "NORMALIZE_WHITESPACE", "IGNORE_EXCEPTION_DETAIL"]
addopts = [
    "--strict-markers",
    "--tb=short",
    "--doctest-modules",
    "--doctest-continue-on-failure",
]

[tool.coverage.run]
source = ["tests"]

[coverage.paths]
source = "embedbase-core"

[coverage.run]
branch = true

[coverage.report]
fail_under = 50
show_missing = true
